Send systemd notifications properly
There's a bug in `systemd-notify` such that the command sends the message and then exits. Since its lifetime is so short, systemd doesn't have time to do the housekeeping to figure out which service unit to associate it with. The mitigation is to just open the socket and send the message directly ourselves, since our process is long-lived.

See https://github.com/systemd/systemd/issues/2737

Fixes #16